test: accelerate sqlite state cli tests
